sql: add logictest for CCL code
Simple tests for CCL-only statements in OSS logic test.
Brain-dead test in CCL logictest just to show it works.

Release note: None
