+++ pkg/sql/logictest/testdata/logic_test/ccl
@@ -1,0 +1,9 @@
+# LogicTest: default
+
+# CCL-only statements error out trying to handle the parsed statements.
+
+statement error pq: unknown statement type: \*tree\.Backup
+BACKUP foo TO "bar"
+
+statement error pq: unknown statement type: \*tree\.Restore
+RESTORE foo FROM "bar"
